Move round selection from dev drawer into View Round JSON modal
